requirements.txt
xgboost>=2.0
lightgbm>=4.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.26
pyarrow>=14.0
matplotlib>=3.8
pytest>=7.4
pytest-asyncio>=0.23
scipy>=1.11
websockets>=12.0
cryptography>=41.0
httpx>=0.27
python-dotenv>=1.0
fastapi>=0.111
uvicorn[standard]>=0.30
aiosqlite>=0.20
